requirements.txt
aiodns==4.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.0
attrs==25.4.0
Bottleneck==1.4.2
ccxt==4.5.36
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coincurve==21.0.0
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
docker==7.1.0
executing==2.2.1
fonttools==4.61.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
iniconfig==2.3.0
ipykernel==6.31.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.3
json5==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lean==1.0.223
lxml==6.0.2
markdown-it-py==4.0.0
matplotlib==3.10.8
matplotlib-inline==0.2.1
mcp==1.26.0
mdurl==0.1.2
multidict==6.7.1
nest-asyncio==1.6.0
numexpr==2.14.1
numpy==2.4.1
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
pip==25.3
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycares==5.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.13.1
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.2.5
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
pyzmq==27.1.0
quantconnect-stubs==17512
referencing==0.37.0
requests==2.32.5
rich==14.3.2
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
setuptools==80.10.2
six==1.17.0
sse-starlette==3.2.0
stack_data==0.6.3
starlette==0.52.1
threadpoolctl==3.5.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
wcwidth==0.2.14
wheel==0.46.3
yarl==1.22.0
